+++ packages/lib-react-components/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased] 2023-12-15
+## [1.12.0] 2024-03-04
 
 ### Added
 Added an `onSignIn` function to AuthModal. A handler than can be called in the parent app when a user registers or signs-in.